Update work
Start working on Vulkan renderer.
Currently the CMake build system is a bloody mess, but it works, at least for linux.
Linking to Vulkan libraries is not yet complete, still need to do that.
A bunch of IDE specific files have been added because I didnt update the gitignore fast enough, however they will be removed in subsequent updates.